[workspace model] provide implementation of FacetManager which stores data in workspace model
Configuration of each Facet is stored in FacetEntity, changes made via FacetManager are propagated to WorkspaceModel and vice versa. Shared code in FacetManager's implementation is extracted for FacetManagerBase. In order to properly reflect changes in configurations of individual facets, a new method FacetManager::facetConfigurationChanged is introduced which must be called after changing configuration of a facet.

GitOrigin-RevId: 0e866dcc7f5769a810bde17cbdd20e4200dfbd09