Copy some files before Lint IDE module extraction
...to preserve git blame for important
files like LintIdeClient and LintIdeProject.

With this change in git history, the Lint IDE module extraction change
adds ~1200 fewer lines, which implies that we preserve ~1200 more lines
of git blame. I've verified that git blame works as expected for the
files involved.

I've used `git rev-parse HEAD:` to verify that the final tree
is the same with and without this change.

Change-Id: I155f38409fd5d7e63ed47872d611d18301be8c17