[Gradle Sync] Remove recommended update scheduling
This change removes the scheduling of notifications if one day has not
yet passed. Instead the user will have to re-sync to see the update. Or
at least this is the intended behaviour.

As far as I can tell this scheduling can never happen since the "next
reminder time" seems to be the last one which will mean current >= next
is always true.

Bug: N/A
Test: N/A
Change-Id: Ieda0203985a840cd509e1d894c09b78629d3830b