Add sequential rendering function in DesignSurface
The sequential rendering is also used by default in DesignSurface.
Before the render requests are not started sequentially, but layoutlib
doesn't support parellel rendering so the renderings are still
executed sequentially. Thus the behaviour is same as before.

Although the behaviour is same as before, but it gives the chance to
interrupt the queued render requests in the following CLs.

Bug: 143308628
Test: manually
Change-Id: Ia74a7e672ce66928b6e26c33197e076ddf00508f