[Dsl] commit documents after save if necessary
Saving a file can modify the associated document: for example, as
mentioned in the documentation for saveDocument(), 'Strip trailing
spaces on Save' can alter the document.  Without committing at this
point, future modifications run the risk of operating on an
Uncommitted document, leading to consistency issues or failures.  The
particular symptom of this in the test suite was a failure on
KotlinScript files with multiple parse/applyChanges cycles: eventually
one had a trailing space, which was removed, but a subsequent
operation used the Psi from the stale document.

Bug: n/a
Test: existing
Change-Id: Ic7811a3a7efd79ae41746ccf4da0f55f3ee55381