Add RunConfiguration to launch Compose Previews to a device
Create ComposePreviewRunConfigurationType and ComposePreviewRunConfiguration
to represent configurations to launch the Compose tooling PreviewActivity to
a device/emulator passing a @Composable to be set as its main content.

This allows running Compose Previews directly on a device/emulator, speeding
up the process of checking how your @Composable's will look like when used
in specific targets.

We don't register ComposePreviewRunConfigurationType in compose-designer.xml
yet, otherwise it would show up in the list of RunConfiguration. That should
only happen when the feature is ready to be enabled.

Bug: 144500928
Test: ComposePreviewRunConfigurationTest
Change-Id: Ic6ec6cb9b4848897e030a08f70c54393ac6ea7fc