Dispose AppInspectionHostService and AppInspectionDiscovery
When disposed (ex: when ide is closed), AppInspectionHostService and AppInspectionDiscovery will perform some basic cleanup.

I suspect this cleanup is not needed because AppInspectionHostService is an Application level service, and the only time it is disposed is when the ide is closed/restarted. However, it is nice to have for writing tests.

Test: AppInspectionDiscoveryManagerTest
Bug: 147321944
Change-Id: Ib1d53c7eb67c3201f9f9dfe8abadeaac30907902