Add full-project highlighting benchmark
The benchmark runs syntax highlighting on all files in the project,
then sets up a separate Perfgate trendline for each file type.
Since timing samples are reported on a per-file basis, Perfgate will
be able to show us the full distribution of file highlighting
times---which is important for getting insight into performance
on large/complex files.

The benchmark also logs diagnostic information, such as the list of
analyzed files and their corresponding highlighting time.

So far this adds a benchmark only for the SantaTracker project.
The test takes around 5 minutes to run locally, most of which is
due to Gradle sync.

Bug: 144098070
Test: //tools/adt/idea/ide-perf-tests/...
Change-Id: I7ab4b4fa56a5c446abf1bf37b5396d8b4fd8a37b