[Gradle Sync] Switch to setting up dependencies via datanodes.
This change removes DependenciesAndroidModuleStep from both
AndroidModuleSetup and AndroidVariantChangeModuleSetup.

Instead we have implemented similar functionality in DependencyUtil.kt
which operates on a ModuleData data node and creates children to
represent all dependencies that were obtained from the
IdeVariant.

We also re-use this method to replace the setup code in
BuildVariantUpdater.java to ensure dependencies are updated on a variant
change. Note: In order to correctly preserve content roots we have had to
extend the scope of modules operated on by the BuildVariantUpdater from
those modules affected by the variant change to all modules in the
project.

Test: tests updated
Fixes: 125680482
Fixes: 135453395
Change-Id: I6051a85ba74cba3dd35be8f5575cfcb757e15f90