Fix split editor not respecting default mode
We were trying to restore the surface state in DesignToolsSplitEditor
too soon, which was resulting in ignoring the default layout defined in
the settings panel. Instead, we were opening new files in the same mode
the last resource file was open with.

This CL changes DesignToolsSplitEditor#getComponent() to first trigger
the ancestor classes' logic, which will select the action corresponding
to the default layout. Afterwards, we can properly set the design surface
state based on the selected action.

Fixes: 147506596
Test: Manual. Newly created files should be open in the default mode
      defined in the settings panel, not according to the last file
      that was open.

Change-Id: I66df58bd7885a6a9e8fb0d14bb188c73f668b123