[Sync] check for Gradle settings as well as build files
Gradle preferentially looks for settings.gradle[.kts] files before
build.gradle[.kts] files to initialize and configure a project.  This
means that users can set things up to have only a settings file in the
root of the project, with all other build scripts kept apart.
However, we were attempting to guess whether an as-yet unimported
project should be treated as a Gradle-based project based on the
presence of build.gradle files only.

Add a check for settings files in the root of the project for when we
are attempting to detect what kind of project we have.

Bug: 147246554
Test: added to GradleProjectInfoTest
Change-Id: Id5a8cfdbb1ef34780ef05512305fcb75b3d5b19b