Allow to download missing or new material icons
Guarded under new StudioFlag.

Downloads the newest metadata file, then it figures out which icons it
needs to download, based on whether there are new icons in this file or
any of the icons have a higher version.

Downloads the necessary icons incrementally, so it downloads all
available styles for an icon, and it updates the existing metadata file,
so that if the download process is interupted it can continue where it
left off.

Bug: 141628234
Test: MaterialIconsDownloaderTest
Change-Id: I6808da2c4999a1b7356238069f32ddb1bfdff7d6