Wait for background tasks to finish after creation
This will ensure the IDE is in a consistent state before proceeding.
It avoids the Layout Editor having to wait to sync or other tasks when
disposing.

In NewProjectTest.testGradleWrapperIsExecutable, the project is closed
before the background tasks are done. Because of this, the Layout Editor
might be waiting and will take longer than usual to dispose.

Fixes: 147416581
Test: Fixes test
Change-Id: I5cbc4b7314e1061fa3a8d836a7927cf7e1288c36