[Dsl] rename getBlockElement to getPropertiesElement
Although getBlockElement is mostly used in the parsers to retrieve
objects corresponding to blocks from the Dsl structure, the name is
possibly confusing in that Dsl "blocks" are not all implemented as
generalized instances of GradleDslBlockElement (some are maps).

Also, parsing Dsl of the form `dotted.reference.to.map.key` will end
up calling this function, where the underlying object is truly a map
and not a Dsl block at all.

Bug: 144679123
Test: existing (no change in functionality)
Change-Id: I7f98d65823e57f0daa6ae1fc575073092bcf2edf