[KTS] Re-consider extra properties syntax logic.
We adopted array expression to be the default syntax for extra
properties creation, which is fine, but this introduced a very tricky
bug.
We use "ext." prefix when parsing extra properties declared using array
expression, and this is now used at the dsl layer stage to propertly
modify the property according to its syntax.
But, when creating a property from the dsl model, it will be applied to
the file using array expression syntax, and unless the file is reparsed,
the property name won't reflect it's syntax (i.e. "ext."). What can
happen is that, if we refer to that property before reparsing the file,
we won't be able to recognise how to apply the reference to kts file
correctly, and this will break the project.
The newly added tests should explain the case more.

Bug: 148769031
Test: new added for KTS to strenghten the checks.
Change-Id: I3c986372ce5fae8a557881c75ad4c3df2f6745d0
(cherry picked from commit 0249aeea0158e6032f77ab4ff26f5a370227b290)