Propagate a new id value to the MTag
The id attribute is updated using the RenameProcessor which writes
directly to the XmlTag. In this CL we detect when the this happens
and update the MTag right after the rename.

This change makes it possible to later select the same MTag after
the MTag hierarchy is rebuilt after an resource change is detected.

Fixes: 147511820
Test: Unit test added.
Change-Id: I1376cd5bf6ed67284a7d26b1c24a81a4bcad425c