Fix CME that would occur in TabularLayout
Swing components / layouts should ideally only be accessed from the EDT
thread. We usually do this just fine in production, but we actually have
been a bit lazy about that in tests, where we often create / query
components on the test thread.

In this CL, I update TabularLayout to complain if accessed outside of
the EDT thread. I then fixed all tests to only run on the EDT thread, so
that the behavior more closely matches that of production.

Fixes: 147606035
Test: Updated
Change-Id: I90c95154aaf4d7beb36fbcd3ec442bd1d80fb12a