[Dsl,.kts] write using accessor for classpath dependency
The ScriptHandlerScope provided by buildscript { ... } installs an
extension function for DependencyHandlerScope.classpath(), meaning
that within a buildscript dependencies block classpath() can be used
as a configuration accessor (not requiring the string form).

Since the default dependencies block (for a Gradle project, rather
than for the build script itself) does not have a classpath
configuration at all, it's unlikely to help the incautious user to
emit an erroneous classpath configuration as quoted rather than
unquoted -- neither will actually work, and at least the unquoted one
will be highlighted in red by the editor as obviously wrong.

Bug: 138277630
Bug: 142114586
Bug: 147592143
Test: strengthened buildscript test and run for .kts
Change-Id: Ifff094e38870c04a8bd06891c532652ec32cef6d