Add support for remapping inline source code refs
When the stacktrace is generated by Compose, it contains line numbers
that do not match the actual source code. The reason for that is that,
when using inline methods, the trace will point to a secondary data
structure within the class (SMAP) that allows the debugger to correctly
map the lines even when the code does not match the original source.
This change replicates the same steps the debugger does to remap the
lines correctly.

The CL also uses a structure for ViewInfo that now is closer to the
actual implementation on the ui-tooling side.

Bug: 147794531
Test: Adds new test
Change-Id: I1971b099a5b297075f4a6e77f0635c2b3bdc79df