Fix table navigation in inspector properties panel
Mark rows that cannot be expanded for read only in the table model.
This will allow the existing code navigate into a table with only
read only items.

Also fix the backwards navigation into a table.
When the table itself receives focus on a backwards move the code
used to delegate to the next component in the focus traversal
policy. This did not work because the table is before its child
components in the natural order.
Instead request focus on the first editor found from the bottom
of the table.

Fixes: 147601697
Test: Added tests for navigating to read only table.
Change-Id: Ib7d040a75f6af96e8c1017ec22098bcbdea5c599