[KTS] Fix variable reference resolution.
When setting a reference to a property, ensure this
reference is injected using the correct syntax.
We used to consider the referenceText directly for Groovy, which doesn't
work for KTS.
fo injections, we don't have to use type casting (it doesn't break the
project, but it doesn't look good). This is planned for an upcoming
change.

Bug: 147796130
Test: Verified manually through PSD
Change-Id: Ieb136f3e7614542b8d7cc3451ab55f34b575fd42
(cherry picked from commit 28b301fb02316efbed283de90588f50611f36873)