Limit the selection in MotionEditor to a few types.
The MotionEditor instance should only keep track of selected:
constraint set, transition, or motion layout.
All other selections are sub selections in other places.

Before this CL a selection in the deisgn surface or the component
tree would set the selected object in the MotionEditor to a
specific constraint. The MotionEditor would then not be able to
restore the selection after a resource change.

Fixes: 147699651
Test: Logged b/147892778 for adding a unit test for this fix.
Change-Id: I4e09ce6e2276aaccdb30982fc2ab06027f1b46aa