[Dsl] improve RepositoriesModel tests
For the setUrlForMethodCall test, have the test input file be of the
form that the test is explicitly testing (i.e. a repositories block
with a zero-argument call to `jcenter()`) rather than programmatically
constructing that test file through the model.

Now that RepositoryClosureTransform is gone, we can also test for
adding artifactUrls to a zero-argument method call repository
declaration.

Bug: 144679123
Test: new in this CL
Change-Id: I562644fe484ff7564056db5e5b33e8d8249cfb75