New DSL: Convert BaseConfig etc. to kt
While this is a binary-compatible change, it is not a source-compatible
one. Update the usages in Studio code and tests.

This implied fixing the type of versionCode. It is of type Int?, and it
only happened to work before in groovy and the result would be a version
code of the character code of the char.

Bug: 140406102
Test: No new tests, minor refactor to existing code.
Change-Id: I03e04b0a4acac4a2cdba78e562af2e98d586c964