Sort the method names for populating OnClick
Before this change the class names and method names given in the
dropdown for the attribute "onClick" would be in an arbitrary order.
If the item count were large that would make it difficult to find
and select the method the user wanted to pick. Sort them to make
this easier. This also fixes a flaky test.

Fixes: 146019765
Test: This is a fix of a flaky test.
Change-Id: I7bee54a0bea3d5c3a87fe0c6b7ddce749e5c6a0d