[jps builds] make ModuleLevelBuilder::getCompilableFileExtensions abstract to ensure that it is implemented
It's important to know which files are processed by ModuleLevelBuilder to skip processing irrelevant files in caches for ModuleBuildTarget. Before the change a notification was printed to the log if some builder doesn't override the method. Now it's will be a compilation error, but binary compatibility isn't broken because we will catch AbstractMethodError at runtime.

GitOrigin-RevId: fdfed347a2904ee31745f7fdf8210b78bc332a97