Initial implementation of AndroidTestSuiteDetailsView
AndroidTestSuiteDetailsView displays detailed results of an instrumentation
test case. The test case may be executed by multiple devices and this view
can show them all. It has a device selector list view at left side and
the result of the selected device is displayed at the right side.

See the buganizer for the spec and design mocks.

Bug: 143299082
Test: unit tests added
Change-Id: I90390f5470b4cf2a1c9dac32a3ba05d7f97758d9