Remove ComposePreviewRunConfigurationProducer from ignored
ComposePreviewRunConfigurationProducer is added to the list of ignored
producers when the COMPOSE_PREVIEW_RUN_CONFIGURATION flag is disabled.
Since RunConfigurationProducerService persists the state, this producer
would remain ignored even if the flag is enabled, as long as it was
ignored at least once. This CL fixes the behavior by removing the
producer from the list of ignored if the flag is enabled.

Bug: 144500928
Test: N/A
Change-Id: I9ebeee37dadb1eb36c04355ff090e76e9d572d9c