[KTS] Convert interpolations to the build language syntax.
When creating an artifact, we accept string values as they are without
checking if they are compatible with the language of the file we are
applying the changes to. Some clients can be aware of the difference and
use the right syntax, but this is not always guaranteed.
This CL adds a check for interpolations in order to convert them to the
corresponding syntax before creating psi elements.

Bug: 147796130
Bug: 147729279
Bug: 148283067
Test: Added

Change-Id: Ic94bd949df098341b6623f105307fe361636c854