[Windows] Ignore line endings in icon generator tests
And also ensure that generated files are written with appropriate line
endings. The images still won't match due to fonts, but this change
fixes file comparisons. Line endings for the golden file can vary due to
Git checkout settings, so convert them to match the project's
CodeStyleSettings, which is what the icon generator should also use.

Bug: 141868211
Test: these are the tests
Change-Id: I60014a18b802593ccaed205909e26c8e48fa7624