[Dsl] support identifiers with dots
Identifiers with dots are particularly problematic with the current
implementation of GradleNameElement, which relies on splitting at dots
to reconstruct hierarchy.  This is lossy in the presence of
identifiers containing dots unless escaping is handled properly and
consistently.

This patch lightly modifies the GradleNameElement representation so
that dots (and backslashes) are escaped on creation from external
sources (though not from GradleNameElement.fake(), which allows
internal use to bypass escaping when we know what we need).  This
escaped notation is isomorphic to a list of string parts, as long as
we unescape when splitting and escape when joining.

When converting external Psi names to GradleNames, handle names
consisting solely of a single quoted string by escaping the string
text.  This isn't complete: a complete solution will need to visit a
dotted name and handle individual parts (and even injections), but
this is enough to support dotted build variant names in the PSD.

Bug: 126937269
Test: added split/join unit tests
Test: added buildType creation and parsing test
Change-Id: Ided4afe287eebedaf2bc90f73ed05b2008d4ddee