Supply explicit Executor for async execution
This addresses this lint warning which now triggers after
recent fixes in lint:

Not specifying an Executor for running callbacks is a common source of threading
issues, resulting in too much work running on the UI thread or the shared
ForkJoinPool used by the IDE for highlighting.

In Android Studio, always specify the executor for running callbacks. As explained
in the documentation for `ListenableFuture.addCallback`, try to avoid
`directExecutor()`.

For more, see `go/do-not-freeze`.

=========

src/com/android/tools/idea/rendering/RenderTask.java:1121 Use `thenComposeAsync` overload with an explicit Executor instead. See `go/do-not-freeze`. [ImplicitExecutor]
    return RenderService.runAsyncRenderAction(() -> measure(modelParser))

Test: N/A
Bug: N/A
Change-Id: Ic34091fe3decec33a17732ce5075cf3e8415b727