[Res Explorer] Support rendering colors for gradle resources
This change removes the restriction to only accept DesignAsset objects
in the rendering pipeline. Since not all resources necessarily require
an external file reference to render.

This is the case for color value resources, since we only need to know
the color value.

Since gradle-defined color resources wouldn't have a
reference to their gradle file they would not have an icon preview even
though we had the color value available. They should now be able to be
previewed in the Resource Manager.

Fixes: 137168924
Test: ColorIconProviderTest
Change-Id: I4605ed1906f969fccea67f549172cb322218ac42