Don't error out computing merged manifest with duplicate activities
If an activity with the same fully-qualified name is declared
multiple times with the same attributes in a module's
manifest, Gradle / the manifest merger will happily
output a merged manifest with duplicate activities (if the
attributes don't match, there will be a build error). In
MergedManifestSnapshotFactory we pre-compute a map from activity
name to attributes based on the merged manifest. At some point
in 3.5, we switched to using an ImmutableMapBuilder instead of
a HashMap, and now the builder is throwing an exception when
it encounters two activities with the same name, even though this
is perfectly valid.

This CL updates MergedManifestSnapshotFactory to use a HashMap
when parsing the merged manifest and eventually create an immutable
copy of it for the end result. The duplicate activities should be
harmless since they have matching attributes (otherwise there'd
be an error from the manifest merger and we'd surface it to the
user appropriately).

Fixes: 139859267
Test: MergedManifestManagerTest#testDuplicateActivities
Change-Id: I875b302a03187dc3121161a088374a88ee60d34d