[Dsl] return ExternalNameInfo in maybeTrimForParent
maybeTrimForParent does two things: it removes name parts from the
hierarchical name of the element, and renames the last part of the
name, both given its parent.  Although it might not be ideal that a
single function does these two things (or maybe that the name of this
function is not a good reflection of what it does), it doesn't change
the nature of the return value, which is of the same kind as the
return value from GradleNameConverter.externalNameForParent() -- that
is, an ExternalNameInfo.

Bug: 126937269
Test: existing (no change in functionality)
Change-Id: Ic6bf7519196557692f6884ee38f82bb5bd5af05f