Enable horizontal scolling when table is wider than viewport.
By default we want all columns of the table to fill the entire width of
the parent. (eg. if there is only one column it should fill the entier
table's width).

If we have enough columns to fill the parent's width and some more,
we want the table to be scrollable horizontally.

Test: TableViewImplTest.kt
Bug: NA
Change-Id: Ic74af142e54ead640b70712d36714e4f3337eddb