Disallow running ComposePreviewRunConfiguration in profiler
ComposePreviewRunConfiguration shouldn't be profilable, so we need to
disallow running it in the Android Profiler. Currently, we're assuming
that every AndroidRunConfigurationBase should be profilable, but we need
to parameterize it so the configuration can decide whether it's profilable.
We default AndroidRunConfigurationBase#isProfilable to true to keep the
current behavior, but override it in ComposePreviewRunConfiguration.

Bug: 144500928
Test: Existing + ComposePreviewRunConfigurationTest
Change-Id: I8cd043ea0dcbbd36b731245833c552e06353c29d