Run CreateNewLibraryModuleWithDefaultsTest in studio_rbe.
CreateNewLibraryModuleWithDefaultsTest runs 10,000 times without
flakes. Its run time is also less than 1 minute. As such we think
this can run as part of the regular presubmit tests in studio_rbe.

This CL removes the qa_sanity tag which triggers this test to
be picked up as part of studio_rbe. We will monitor how this
target behaves in the presubmit tests over the next few weeks.

Bug: 148463664
Test: CreateNewLibraryModuleWithDefaultsTest
Change-Id: Id0ca504731a66eef79ffd09c95bf8a402342d69c