Merge AndroidRunConfiguration with its parent class
AndroidAppRunConfigurationBase is an abstract class with a single child,
so we can merge them. That will improve maintainability, since we won't
need to worry about using AndroidRunConfiguration vs its parent in the
code base (we already have multiple occurrences of "instance of" for
both).
Merging these classes will prevent potential issues caused by
subclassing
AndroidAppRunConfigurationBase in the future when we were supposed to
subclass AndroidRunConfiguration instead.

Bug: N/A
Test: existing tests should pass after refactoring
Change-Id: I0d257cf121508607dfee52c8e5e702ab9379a253