[Dsl,.gr] create names using application syntax
If we create Psi elements for names using a PsiFactory on an
assignment expression, we restrict the scope of possible names.
Specifically, constructing Psi from text of the form
  'foo' = 1
produces a GrAssignmentExpression whose lvalue is a literal, not a
reference expression, which causes problems downstream when attempting
to reparse a GrApplicationStatement after applying changes to the
build file.

Instead, construct Psi from text of the form
  'foo' 1
which is lexically legal and parses as a GrApplicationStatement whose
invoked expression is a GrReferenceExpression.

Bug: 126937269
Test: existing (plus tests in forthcoming CL)
Change-Id: I88ecec15feabd179c1dc7a2f60b16d255e4cd184