Don't call startProfiling() on unmatched devices
When multiple instances of Studio are open, there may be
multiple instances of StudioProfiers class. They all
register to PROFILING_SESSION aspect and
profilingSessionChanged() will be called on each of the
instance. We should not call startProfiling() if the
device isn't the one where the new session is from.

Ideally, we should call startProfiling() only once, but
it's not easy in the current implementation. This CL
fixes an exception in the IDE, but Profiler still doesn't
support running two or more profiler sessions
simultaneously.

Bug: 146037091
Test: manual
Change-Id: Ifb0ff99342c68102ef9a0532ec1479e4f205a2db