Don't show logcat in run toolwindow when instrumentation
AndroidProcessHandler captures logcat message after the
handler finds target processes. Because lifespan of test
application process is typically very short, execution of
AndroidTestRunConfiguration finishes before the handler
finds the process. This causes flakiness and logcat message
is not displayed in the run toolwindow reliably.

To eliminate the flakiness, this CL disables showing logcat
message in run toolwindow if the run configuration type is
AndroidTestRunConfigurationType.

(For long-term, we will revive the logcat message in new test
suite panel with Nitrogen test executor.)

Fixes: 148217065
Test: unit tests added
Change-Id: I2e5148789f07e16c33663e8dc6dba7e042eabd55