Provide the best results when parsing manifest file
... for AndroidManifestIndex. At least we output everything that's
parsed so far until we hit the syntax issue in the manifest. For
the rest of manifest file, we try to parse as much as we can with
correct tag matching guaranteed. XmlPullParserException now are thrown
and caught silently as much as possible.

Bug: 142260247
Test: AndroidManifestIndexTest
Change-Id: Ib642e6e03a5c99bc88cd0eae5044acc00d776ef9
(cherry picked from commit 42b9922832a8e877a78d954463afc22e49419fd8)